requirements.txt
torch>=1.10.0
torchvision>=0.11.1
Pillow
opencv-python
numpy
transformers>=4.15.0
ultralytics>=8.0.0
timm==0.4.12
einops
fairscale
yacs
tqdm
